# docs/openapi/ecom-api.yaml
openapi: 3.0.3
info:
  title: EcomAI API (Phase 1)
  version: 0.1.0
  description: |
    管理端 (8080) 与用户端 (8085) 契约。响应格式统一为 mall CommonResult。
servers:
  - url: http://localhost:8080
    description: Admin API
  - url: http://localhost:8085
    description: Portal API

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    CommonResult:
      type: object
      properties:
        code: { type: integer, example: 200 }
        message: { type: string }
        data: {}
    LoginParam:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    RegisterParam:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
        nickname: { type: string }
        telephone: { type: string }

paths:
  /admin/login:
    post:
      tags: [admin]
      summary: 管理员登录
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginParam' }
      responses:
        '200': { description: token }
  /admin/info:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 用户信息与菜单 }
  /product/list:
    get:
      tags: [admin-product]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: pageNum, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: publishStatus, in: query, schema: { type: integer } }
      responses:
        '200': { description: 商品分页 }
  /product/update/publishStatus:
    post:
      tags: [admin-product]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: ids, in: query, schema: { type: array, items: { type: integer } } }
        - { name: publishStatus, in: query, schema: { type: integer } }
      responses:
        '200': { description: OK }
  /order/list:
    get:
      tags: [order]
      parameters:
        - { name: status, in: query, schema: { type: integer } }
        - { name: pageNum, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
      responses:
        '200': { description: 订单列表 }
  /sso/register:
    post:
      tags: [portal]
      summary: 注册
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterParam' }
      responses:
        '200': { description: 注册成功 }
  /sso/login:
    post:
      tags: [portal]
      summary: 登录
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginParam' }
      responses:
        '200': { description: token }
  /home/content:
    get:
      tags: [portal]
      summary: 首页聚合
      responses:
        '200': { description: OK }
  /product/detail/{id}:
    get:
      tags: [portal]
      summary: 商品详情
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: OK }
  /cart/add:
    post:
      tags: [portal]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /order/generateOrder:
    post:
      tags: [portal]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 下单成功 }
  /order/paySuccess:
    post:
      tags: [portal]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 支付成功 }